fix joinAll()
It can happen, that if joinAll() is called twice in a short amount of time, that the code stuck inside the while loop of

decWorkerCountBy(std::numeric_limits<size_t>::max(), Method::SYNC);

That's because not all threads are actually joined when the second joinAll() is called.

I solved the problem by adding while(getWorkerCount() > 0) { } at the end of joinAll(), however that's probably not the best solution.